:sparkles: Added UserDao, UniversityDao and InsertPortoMetroSystemSql
Implemented a findAllUsers() method in the MySqlUserDao class which returns a list of all Users from the database.
However, it does not return the card details yet.

Implemented a findUniversityByUniversityId(String universityId) method in the MySqlUniversityDao class which returns a University for a Student user.

Implemented a InsertPortoMetroSystemSql class which runs the porto_metro_system.sql in MySQL XAMMP.