:recycle: Refactor ControllersUtil to make it easier to use
Removed template from ControllersUtil as it actually results to more duplicate in codes since it needs to be initialised everytime a redirect is needed. Now every page has a redirect function in ControllersUtil. This makes it easier to call a redirect, as it takes fewer parameters and only needs to be initialised once in every controller.