:sparkles: Add a ScheduleDao and TimetableDao and BasicDataSource
When creating a ScheduleDao, a stations Hashmap<String, Station> is used to keep track of stations.

Instead of fetching the StationDao for a station object and creating a new instance, the CreateDao() just retrieve it from the hashmap if the station already exists.

This decreased the schedule fetching time from 5100ms to 21ms. It even goes as fast as 9 ms when all stations are in the Hashmap.

BasicDataSource keeps a pool connection rather than creating a new connection every time which can be costly.