:hammer: Finish ImportTimetables class
Timetable can be inserted and retrieved.

Add ComparatorSchedules to compare Schedule by there station in respect to the index of the stations set for that timetable.

This class is needed to so that when Schedules are imported the times are ordered by the Station position in the Timetable and not by the LocalDateTime of the Schedule.