:sparkles: Add setTimetableRowOrder and getClosestScheduleToStartTime
The setTimetableRowOrder method uses the ComparatorSchedules class to order each schedule by the order of the schedulesByStation LinkedHashMap Station key.

The getClosestScheduleToStartTime method uses a helper method checkEndScheduleInvalid to remove any schedules that are on that column that contain the end station and the end invalid time. This stops the Passenger being able to pick a valid schedule column but the end station has a invalid time.

The getClosestScheduleToStartTime uses a PriorityQueue and the ComparatorSchedulesClosestTime comparator to return the peek.